Move DomainCheckJob under Jobs\RequestSSL
